# public/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "boiler_api_5300",
		"version": "1.0.0",
		"description": "Boilerplate apis",
		"license": {
			"name": "license"
		},
		"contact": {
			"name": "STS"
		}
	},
	"paths": {
...
		"/user/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get User login",
				"tags": [
					"User Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Save a User",
				"tags": [
					"User Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"Name": {
										"type": "string"
									},
									"Email": {
										"type": "string"
									},
									"phone_no": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"Role": {
										"type": "string"
									},
									"Address": {
										"type": "string"
									},
									"profile_pic": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"Name",
									"Email",
									"phone_no",
									"password"
								]
							}
						}
					}
				}
			}
		},
		"/user/auth/forgot_password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Forgot password api endpoint",
				"tags": [
					"User Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/auth/reset_password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Reset password api endpoint",
				"tags": [
					"User Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"otp": {
										"type": "string"
									},
									"new_password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"otp",
									"new_password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/auth/verify_otp": {
			"post": {
				"operationId": "VerifyOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Verify Otp Route  api endpoint",
				"tags": [
					"User Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"otp": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"otp",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/auth/resend_otp": {
			"post": {
				"operationId": "ResendOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Resend Otp Route  api endpoint",
				"tags": [
					"User Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/auth/change_password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Change Password endpoint",
				"tags": [
					"User Auth Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"new_password": {
										"type": "string"
									},
									"old_password": {
										"type": "string"
									}
								},
								"required": [
									"new_password",
									"old_password"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/auth/details": {
			"get": {
				"operationId": "GetUserDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get User info",
				"tags": [
					"User Auth Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		},
		"/user/auth/profile": {
			"put": {
				"operationId": "UpdateUserProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update User Profile",
				"tags": [
					"User Auth Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"Name": {
										"type": "string"
									},
									"phone_no": {
										"type": "string"
									},
									"Address": {
										"type": "string"
									},
									"profile_pic": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		},
		"/user/auth/delete_deactivate": {
			"delete": {
				"operationId": "DeleteOrDeactivateAccount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "delete or deactivate user account\r\n2 for delete 3 for deactivate",
				"tags": [
					"User Auth Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"reason": {
										"type": "string"
									},
									"status": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"status"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/auth/refresh_token": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Refresh tokne api\r\nprovide refresh token in this api and get new access token",
				"tags": [
					"User Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"refresh_token": {
										"type": "string"
									}
								},
								"required": [
									"refresh_token"
								]
							}
						}
					}
				}
			}
		},
...
		"/user/auth/upload_file": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Upload a file",
				"tags": [
					"User Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/user/auth/details_user": {
			"get": {
				"operationId": "GetUserDetailsUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get User info",
				"tags": [
					"User Auth Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		},
...
		"/admin/user/details": {
			"get": {
				"operationId": "GetUserDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get User Details",
				"tags": [
					"Admin User Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/admin/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get Admin login",
				"tags": [
					"Admin Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/auth/forgot_password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Forgot password api endpoint",
				"tags": [
					"Admin Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/auth/reset_password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Reset password api endpoint",
				"tags": [
					"Admin Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"otp": {
										"type": "string"
									},
									"new_password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"otp",
									"new_password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/auth/verify_otp": {
			"post": {
				"operationId": "VerifyOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Verify Otp Route  api endpoint",
				"tags": [
					"Admin Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"otp": {
										"type": "number",
										"format": "double"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"otp",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/auth/resend_otp": {
			"post": {
				"operationId": "ResendOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Resend Otp Route  api endpoint",
				"tags": [
					"Admin Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/auth/change_password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Change Password endpoint",
				"tags": [
					"Admin Auth Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"new_password": {
										"type": "string"
									},
									"old_password": {
										"type": "string"
									}
								},
								"required": [
									"new_password",
									"old_password"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/auth/details": {
			"get": {
				"operationId": "GetAdminDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get Admin info",
				"tags": [
					"Admin Auth Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		},
		"/admin/auth/profile": {
			"put": {
				"operationId": "UpdateAdminProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update Admin Profile",
				"tags": [
					"Admin Auth Routes"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"first_name": {
										"type": "string"
									},
									"last_name": {
										"type": "string"
									},
									"phone_number": {
										"type": "string"
									},
									"country_code": {
										"type": "string"
									},
									"greet_msg": {
										"type": "string"
									},
									"profile_pic": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		},
		"/admin/auth/refresh_token": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Refresh token api\r\nprovide refresh token in this api and get new access token",
				"tags": [
					"Admin Auth Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"refresh_token": {
										"type": "string"
									}
								},
								"required": [
									"refresh_token"
								]
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}